Add files via upload
Here is some code that hopefully is how I would create a very basic neural network to train the data. The last thing I need to do is get my labels or y values. According to the GW website, the labels should be stored in the DQmask. At the bottom I'll link where I'm finding the description on the DQmask and quality data. When I run the code and have it print out the DQmask, it seems to give me different values than what the link below tells me should be in it.
https://www.gw-openscience.org/tutorial03/